+++ core/src/main/scala/zio/model/Cookie.scala
@@ -1,4 +1,4 @@
-package zio.http
+package zio.model
 
 import java.time.LocalDateTime
